% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_data.R
\name{adjust_age}
\alias{adjust_age}
\title{Utility function: adjust age to reflect relevant seasons}
\usage{
adjust_age(df)
}
\arguments{
\item{df}{the dataframe, which must contain an age and season column, in which to apply the adjustments}
}
\value{
the dataframe with age adjusted
}
\description{
Utility function: adjust age to reflect relevant seasons
}
\examples{
adjust_age(df)
}
