% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{compute_next_year_fantasy_pts}
\alias{compute_next_year_fantasy_pts}
\title{Append next-year fantasy points (PPR and regular) to existing table containing present-year info}
\usage{
compute_next_year_fantasy_pts(joined_dfs)
}
\arguments{
\item{joined}{a dataframe. It can be the output of join_weekly_and_seasonal but just needs to include fantasy point and games data}
}
\value{
a dataframe with present- and next-year fantasy point and games stats
}
\description{
Append next-year fantasy points (PPR and regular) to existing table containing present-year info
}
\examples{
compute_next_year_fantasy_pts(joined)
}
