% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{find_id}
\alias{find_id}
\title{Utility function: via INNER JOIN, lookup gsis_id given some other valid id type.}
\usage{
find_id(df, id_type = "pfr_id")
}
\arguments{
\item{df}{the dataframe in which to populate gsis_id via inner join with lookup}

\item{id_type}{a string indicating the name of the id_type (besides gsis_id) to use in the lookup.}
}
\value{
the seasonal summary dataframe for snap counts
}
\description{
Utility function: via INNER JOIN, lookup gsis_id given some other valid id type.
}
\examples{
find_id(df, "nfl_id")
}
