% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{get_data}
\alias{get_data}
\title{Pull conditioned data for specified seasons}
\usage{
get_data(min_year, max_year)
}
\arguments{
\item{min_year}{an integer year correpsonding to the earliest valid season of interest}

\item{max_year}{an integer year correpsonding to the latest valid season of interest}
}
\value{
a conditioned dataframe of stats at the player-season level that can be used for training
}
\description{
Pull conditioned data for specified seasons
}
\examples{
load_weekly()
}
