% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_data.R
\name{get_player_bios}
\alias{get_player_bios}
\title{Get player bios from ff_playerids dictionary of nflreadr}
\usage{
get_player_bios()
}
\value{
player bios table w/ gsis_id as only player ID
}
\description{
Get player bios from ff_playerids dictionary of nflreadr
}
\examples{
get_player_bios()
}
