% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_blank_median}
\alias{impute_blank_median}
\title{Impute blank for character variables and column median for numeric variables}
\usage{
impute_blank_median(df)
}
\arguments{
\item{df}{the dataframe on which to perform the imputaiton}
}
\value{
a dataframe with the values imputed
}
\description{
Impute blank for character variables and column median for numeric variables
}
\examples{
impute_blank_median(df)
}
