% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_data.R
\name{impute_blank_zero}
\alias{impute_blank_zero}
\title{Utility function: impute missing values, replacing with 0 for numeric variables and "" for character variables}
\usage{
impute_blank_zero(df)
}
\arguments{
\item{df}{the dataframe in which to perform the imputation}
}
\value{
a dataframe with all numeric and character columns fully imputed w/o NAs
}
\description{
Utility function: impute missing values, replacing with 0 for numeric variables and "" for character variables
}
\examples{
impute_blank_zero(df)
}
