% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_player_bios}
\alias{join_player_bios}
\title{Utility function: joins datframe w/ gsis-type player_id and season to player bios data}
\usage{
join_player_bios(dfs)
}
\arguments{
\item{dfs}{the dataframe, w/ gsis-style player_id and season as columns, to inner join with player bio info}
}
\value{
dataframe with player bio info
}
\description{
Utility function: joins datframe w/ gsis-type player_id and season to player bios data
}
\examples{
join_player_bios(df)
}
