% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{join_weekly_and_seasonal}
\alias{join_weekly_and_seasonal}
\title{Join weekly and seasonal data together into single table}
\usage{
join_weekly_and_seasonal(weekly, seasonal)
}
\arguments{
\item{weekly}{the dataframe of weekly data from load_weekly()}

\item{seasonal}{the dataframe of seasonal data from load_seasonal(x, y)}
}
\value{
a dataframe with joined weekly and seasonal data
}
\description{
Join weekly and seasonal data together into single table
}
\examples{
join_weekly_and_seasonal(weekly, seasonal)
}
