% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{left_join_by_player_id_and_season}
\alias{left_join_by_player_id_and_season}
\title{Utility function: performs left join between two dataframes on player_id and season}
\usage{
left_join_by_player_id_and_season(dfA, dfB, suffixA = "", suffixB = "")
}
\arguments{
\item{dfA}{the first dataframe (must contain player_id and season)}

\item{dfB}{the second dataframe (must contain player_id and season)}

\item{suffixA}{the suffix to apply to the left side (defaults to "")}

\item{suffixB}{the suffix to apply to the right side (defaults to "")}
}
\value{
the joined dataframe
}
\description{
Utility function: performs left join between two dataframes on player_id and season
}
\examples{
left_join_by_player_id_and_season(df1, df2)
left_join_by_player_id_and_season(df1, df2, "_left", "right")
}
