% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{left_join_positional_list_by_player_id_and_season}
\alias{left_join_positional_list_by_player_id_and_season}
\title{Utility function: performs left join between across to lists of dataframe vectors on player_id and season}
\usage{
left_join_positional_list_by_player_id_and_season(
  dfAs,
  dfBs,
  suffixA = "",
  suffixB = ""
)
}
\arguments{
\item{dfAs}{the vector of first dataframe (must contain player_id and season)}

\item{dfBs}{the vector of second dataframe (must contain player_id and season)}

\item{suffixA}{the suffix to apply to the left side (defaults to "")}

\item{suffixB}{the suffix to apply to the right side (defaults to "")}
}
\value{
the vector of joined dataframes
}
\description{
Utility function: performs left join between across to lists of dataframe vectors on player_id and season
}
\examples{
left_join_positinoal_list_by_player_id_and_season(df1s, df2s)
left_join_positinoal_list_by_player_id_and_season(df1s, df2s, "_left", "right")
}
