% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_data.R
\name{load_combine}
\alias{load_combine}
\title{Load snap count data from nflreadr}
\usage{
load_combine()
}
\value{
the summary dataframe for combine data
}
\description{
Load snap count data from nflreadr
}
\examples{
load_combine()
}
