% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_data.R
\name{load_espn_qbr}
\alias{load_espn_qbr}
\title{Load ESPN Quarterback Rating data from nflreadr}
\usage{
load_espn_qbr(min_year, max_year)
}
\value{
the seasonal summary dataframe for QBR
}
\description{
Load ESPN Quarterback Rating data from nflreadr
}
\examples{
load_espn_qbr()
}
