% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{load_pfr_advanced}
\alias{load_pfr_advanced}
\title{Load weekly pfr advanced data from nflreadr. This function is weekly, not seasonal. Only applicable to QBs.}
\usage{
load_pfr_advanced(min_year, max_year)
}
\value{
the weekly pivoted dataframe for snap counts
}
\description{
Load weekly pfr advanced data from nflreadr. This function is weekly, not seasonal. Only applicable to QBs.
}
\examples{
load_pfr_advanced()
}
