% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{load_seasonal}
\alias{load_seasonal}
\title{Pull seasonal stats using nflfastR}
\usage{
load_seasonal(min_year, max_year)
}
\arguments{
\item{min_year}{a number indicating the earliest season (inclusive)}

\item{max_year}{a number indicating the latest season (also inclusive)}
}
\value{
a dataframe of seasonal stats
}
\description{
Pull seasonal stats using nflfastR
}
\examples{
load_seasonal(1999, 2022)
load_seasonal(2017, 2017)
}
