% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_data.R
\name{load_snap_counts}
\alias{load_snap_counts}
\title{Load snap count data from nflreadr}
\usage{
load_snap_counts()
}
\value{
the seasonal summary dataframe for snap counts
}
\description{
Load snap count data from nflreadr
}
\examples{
load_snap_counts()
}
