% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy.R
\name{load_weekly}
\alias{load_weekly}
\title{Pull weekly performance data using nflfastR}
\usage{
load_weekly(min_year, max_year)
}
\value{
a dataframe of weekly stats
}
\description{
Pull weekly performance data using nflfastR
}
\examples{
load_weekly()
}
